Split up components into separate files and added eslint to project
