+++ .eslintrc.js
@@ -1,0 +1,9 @@
+module.exports = {
+  "extends": "airbnb",
+  "rules": {
+    "curly": "off",
+    "no-bitwise": "off",
+    "no-plusplus": "off",
+    "react/jsx-filename-extension": "off"
+  }
+};
